# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transcriptions/labelled": {
      "get": {
        "tags": [
          "Transcriptions"
        ],
        "summary": "List Labelled Transcriptions",
        "description": "Get paginated metadata for labelled transcriptions for the current user.",
        "operationId": "list_labelled_transcriptions_transcriptions_labelled_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (starts from 1)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (starts from 1)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of items per page"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TranscriptionSortOrder",
              "description": "Sort order for date recorded",
              "default": "newest"
            },
            "description": "Sort order for date recorded"
          },
          {
            "name": "client_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by client name",
              "title": "Client Name"
            },
            "description": "Filter by client name"
          },
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by case ID",
              "title": "Case Id"
            },
            "description": "Filter by case ID"
          },
          {
            "name": "subject",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subject",
              "title": "Subject"
            },
            "description": "Filter by subject"
          },
          {
            "name": "date_of_recording",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date recorded",
              "title": "Date Of Recording"
            },
            "description": "Filter by date recorded"
          },
          {
            "name": "date_of_recording_day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 31,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by recorded day",
              "title": "Date Of Recording Day"
            },
            "description": "Filter by recorded day"
          },
          {
            "name": "date_of_recording_month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by recorded month",
              "title": "Date Of Recording Month"
            },
            "description": "Filter by recorded month"
          },
          {
            "name": "date_of_recording_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by recorded year",
              "title": "Date Of Recording Year"
            },
            "description": "Filter by recorded year"
          },
          {
            "name": "client_date_of_birth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by client date of birth",
              "title": "Client Date Of Birth"
            },
            "description": "Filter by client date of birth"
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelledTranscriptionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/recordings": {
      "post": {
        "tags": [
          "Transcriptions"
        ],
        "summary": "Create Recording",
        "operationId": "create_recording_recordings_post",
        "parameters": [
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transcriptions": {
      "post": {
        "tags": [
          "Transcriptions"
        ],
        "summary": "Create Transcription",
        "description": "Start a transcription job.",
        "operationId": "create_transcription_transcriptions_post",
        "parameters": [
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/transcriptions/{transcription_id}/recordings": {
      "get": {
        "tags": [
          "Transcriptions"
        ],
        "summary": "Get Recordings For Transcription",
        "operationId": "get_recordings_for_transcription_transcriptions__transcription_id__recordings_get",
        "parameters": [
          {
            "name": "transcription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transcription Id"
            }
          },
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SingleRecording"
                  },
                  "title": "Response Get Recordings For Transcription Transcriptions  Transcription Id  Recordings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelledTranscriptionsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LabelledTranscriptionMetadata"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "LabelledTranscriptionsResponse",
        "description": "Response for labelled transcriptions."
      },
...
      "RecordingCreateRequest": {
        "properties": {
          "file_extension": {
            "type": "string",
            "title": "File Extension"
          },
          "file_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Created At"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecordingSource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "file_extension"
        ],
        "title": "RecordingCreateRequest"
      },
      "RecordingCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "upload_url"
        ],
        "title": "RecordingCreateResponse"
      },
...
      "TranscriptionCreateRequest": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "recording_id"
        ],
        "title": "TranscriptionCreateRequest"
      },
...
      "TranscriptionSortOrder": {
        "type": "string",
        "enum": [
          "newest",
          "oldest"
        ],
        "title": "TranscriptionSortOrder"
      },
...
    }
  }
}